only use usize for window_size arg

+++ 2021/src/main.rs
@@ -9,7 +9,7 @@ fn main() {
         read_to_string("input/day1_sonar_sweep")
             .unwrap()
             .lines()
-            .map(|r| r.parse::<usize>().unwrap())
+            .map(|r| r.parse::<i32>().unwrap())
             .collect(),
     );
 
@@ -40,7 +40,7 @@ mod tests {
             read_to_string("input/day1_sonar_sweep")
                 .unwrap()
                 .lines()
-                .map(|r| r.parse::<usize>().unwrap())
+                .map(|r| r.parse::<i32>().unwrap())
                 .collect(),
         );
 
@@ -53,7 +53,7 @@ mod tests {
             read_to_string("input/day1_sonar_sweep")
                 .unwrap()
                 .lines()
-                .map(|r| r.parse::<usize>().unwrap())
+                .map(|r| r.parse::<i32>().unwrap())
                 .collect(),
         );
